requirements.txt
Django==2.0.4
django-configurations==2.0
django-rest-framework==0.1.0
